fix : private method 부분 정리
