+++ src/main/java/gdsc/be/mount/storage/Enum/ActionType.java
@@ -1,5 +1,5 @@
 package gdsc.be.mount.storage.Enum;
 
 public enum ActionType {
-    UPLOAD, UPDATE, DELETE, DOWNLOAD
+    UPLOAD, UPDATE, DELETE, DOWNLOAD, READ
 }
